requirements.txt
asgiref==3.8.1
Django==5.1.3
django-jazzmin==3.0.1
django-multiselectfield==0.1.13 #for multiple select
djangorestframework==3.15.2  # for api
sqlparse==0.5.1
tzdata==2024.2
